requirements.txt
transformers>=4.41.0
datasets>=2.19.0
peft>=0.11.1
accelerate>=0.31.0
torch>=2.2.0
pyyaml>=6.0.1
sentencepiece>=0.2.0
